fix: auto-detect repo for issue autopilot
